CLI migrated to .NET Core. Fix some analyzer exceptions Update readme.md
