+++ Linty.CLI/AnalyzerReport.cs
@@ -1,10 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using Microsoft.CodeAnalysis;
-using UnityEngineAnalyzer.CLI.Reporting;
+using Linty.CLI.Reporting;
 using Linty.Analyzers;
 
-namespace UnityEngineAnalyzer.CLI
+namespace Linty.CLI
 {
     public class AnalyzerReport
     {
