+++ Linty.CLI/Properties/launchSettings.json
@@ -1,0 +1,7 @@
+{
+  "profiles": {
+    "Linty.CLI": {
+      "commandName": "Project"
+    }
+  }
+}
