+++ UnityEngineAnalyzer.sln
@@ -11,6 +11,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnityEngineAnalyzer.CLI", "
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Linty.Analyzers", "UnityEngineAnalyzer\Linty.Analyzers\Linty.Analyzers.csproj", "{D43EFDFB-BB7F-4A9B-9A60-C79B9404342A}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Linty.CLI", "Linty.CLI\Linty.CLI.csproj", "{616BA250-393A-4F44-AB95-5C511969B242}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -33,6 +35,10 @@ Global
 		{D43EFDFB-BB7F-4A9B-9A60-C79B9404342A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D43EFDFB-BB7F-4A9B-9A60-C79B9404342A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D43EFDFB-BB7F-4A9B-9A60-C79B9404342A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{616BA250-393A-4F44-AB95-5C511969B242}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{616BA250-393A-4F44-AB95-5C511969B242}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{616BA250-393A-4F44-AB95-5C511969B242}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{616BA250-393A-4F44-AB95-5C511969B242}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
